src/services/API/requirements.txt
fastapi[standard]
sqlalchemy
psycopg[binary]
pyjwt
passlib[bcrypt]
python-dotenv
celery
redis
websockets